# [PEA] At loop header, only mark objects that are not live as escaped