Escape all objects that are not live at loop header. Before we conservatively escape everything. We can strengthen the optimization by only escaping objects that are live at the loop header. We need to mark objects escaped through DFS. It is possible that an object is live, but it has a reference that is not live. That reference is also escaped.

For example, if we inline [HashMap::replace()](https://github.com/openjdk/jdk/blob/009f5e1fa177eea326aefec0f995f589a01169d2/src/java.base/share/classes/java/util/HashMap.java#L1173-L1182), and the value is an object type, i.e. `HashMap<Integer, Object>`, we don’t need the allocate the object is the key is not present. However, if `HashMap::getNode` is inlined, PEA will encounter [a loop](https://github.com/openjdk/jdk/blob/009f5e1fa177eea326aefec0f995f589a01169d2/src/java.base/share/classes/java/util/HashMap.java#L583-L587), and escape everything. `HashMap::getNode` has nothing to do with the value object, and that object is not live at the loop header. With this patch, we can avoid materializing the value object.

If we set `-p fillInterval=1`, PEA does not have any benefits. All objects are used in `HashMap::replace` because the key is present in the HashMap. With the default `-p fillInterval=2`, we avoid the allocation of the object half of the time, since the key is not present and the new object is not needed. PEA reduces 25% allocation rate with default parameters.
Without PEA:
```
Benchmark                                (fillInterval)  (size)  Mode  Cnt        Score         Error   Units
HashMapBench.replace                                  2  100000  avgt   10  3051607.127 ± 1713433.504   ns/op
HashMapBench.replace:gc.alloc.rate                    2  100000  avgt   10     1121.073 ±     585.919  MB/sec
HashMapBench.replace:gc.alloc.rate.norm               2  100000  avgt   10  3197973.058 ±      11.867    B/op
HashMapBench.replace:gc.count                         2  100000  avgt   10        4.000                counts
HashMapBench.replace:gc.time                          2  100000  avgt   10      726.000                    ms
```
With PEA:
```
Benchmark                                (fillInterval)  (size)  Mode  Cnt        Score        Error   Units
HashMapBench.replace                                  2  100000  avgt   10  2624624.627 ± 808835.461   ns/op
HashMapBench.replace:gc.alloc.rate                    2  100000  avgt   10      907.745 ±    319.824  MB/sec
HashMapBench.replace:gc.alloc.rate.norm               2  100000  avgt   10  2397969.855 ±      6.001    B/op
HashMapBench.replace:gc.count                         2  100000  avgt   10        4.000               counts
HashMapBench.replace:gc.time                          2  100000  avgt   10      660.000                   ms
```