@caojoshua 2023-10-03
I pushed a change with `kill_dead_locals()`. It fails our run_streams.sh regression test. I'm not familiar with the API. I was under the impression we should just be able to call it during Parse time.

```
#
# A fatal error has been detected by the Java Runtime Environment:
#
#  Internal Error (/local/home/joshcao/jdk/jdk/src/hotspot/share/opto/type.hpp:2091), pid=42630, tid=42644
#  assert(_base == AryPtr) failed: Not an array pointer
#
# JRE version: OpenJDK Runtime Environment (22.0) (fastdebug build 22-internal-adhoc.joshcao.jdk)
# Java VM: OpenJDK 64-Bit Server VM (fastdebug 22-internal-adhoc.joshcao.jdk, mixed mode, compressed oops, compressed class ptrs, g1 gc, linux-amd64)
# Problematic frame:
# V  [libjvm.so+0x13a226e]  Parse::array_addressing(BasicType, int, Type const*&)+0x4fe
#
# No core dump will be written. Core dumps have been disabled. To enable core dumping, try "ulimit -c unlimited" before starting Java again
#
# An error report file with more information is saved as:
# /local/home/joshcao/jdk/jdk/PEA/hs_err_pid42630.log
#
# Compiler replay data is saved as:
# /local/home/joshcao/jdk/jdk/PEA/replay_pid42630.log
#
# If you would like to submit a bug report, please visit:
#   https://bugreport.java.com/bugreport/crash.jsp
#
run_streams.sh: line 1: 42630 Aborted                 java -Xbatch -XX:-TieredCompilation -XX:+UnlockExperimentalVMOptions $* CountUppercase This year I would like to run ALL languages in one VM.
```

```
Thread 15 "C2 CompilerThre" received signal SIGSEGV, Segmentation fault.
[Switching to Thread 0x7fffb5f5e700 (LWP 44527)]
0x00007ffff62d026e in Type::is_aryptr (this=0x7ffff0376f68) at /local/home/joshcao/jdk/jdk/src/hotspot/share/opto/type.hpp:2091
2091      assert( _base == AryPtr, "Not an array pointer" );
(gdb) bt
#0  0x00007ffff62d026e in Type::is_aryptr (this=0x7ffff0376f68) at /local/home/joshcao/jdk/jdk/src/hotspot/share/opto/type.hpp:2091
#1  Parse::array_addressing (this=this@entry=0x7fffb5f5b9d0, type=type@entry=T_BYTE, vals=vals@entry=0, elemtype=@0x7fffb5f5b708: 0x7ffff0376f68) at /local/home/joshcao/jdk/jdk/src/hotspot/share/opto/parse2.cpp:130
#2  0x00007ffff62d046f in Parse::array_load (this=this@entry=0x7fffb5f5b9d0, bt=bt@entry=T_BYTE) at /local/home/joshcao/jdk/jdk/src/hotspot/share/opto/parse2.cpp:57
#3  0x00007ffff62d954b in Parse::do_one_bytecode (this=this@entry=0x7fffb5f5b9d0) at /local/home/joshcao/jdk/jdk/src/hotspot/share/opto/parse2.cpp:2137
#4  0x00007ffff62c93ab in Parse::do_one_block (this=this@entry=0x7fffb5f5b9d0) at /local/home/joshcao/jdk/jdk/src/hotspot/share/opto/parse1.cpp:1663
#5  0x00007ffff62ca610 in Parse::do_all_blocks (this=this@entry=0x7fffb5f5b9d0) at /local/home/joshcao/jdk/jdk/src/hotspot/share/opto/parse1.cpp:762
#6  0x00007ffff62ceaae in Parse::Parse (this=0x7fffb5f5b9d0, caller=<optimized out>, parse_method=<optimized out>, expected_uses=<optimized out>, caller_state=<optimized out>)
    at /local/home/joshcao/jdk/jdk/src/hotspot/share/opto/parse1.cpp:632
#7  0x00007ffff5724bd4 in ParseGenerator::generate (this=<optimized out>, jvms=0x7fff841c3d38) at /local/home/joshcao/jdk/jdk/src/hotspot/share/opto/callGenerator.cpp:99
#8  0x00007ffff58afc7a in Compile::Compile (this=0x7fffb5f5c850, ci_env=<optimized out>, target=<optimized out>, osr_bci=<optimized out>, options=..., directive=<optimized out>)
    at /local/home/joshcao/jdk/jdk/src/hotspot/share/opto/compile.cpp:788
#9  0x00007ffff5722eb7 in C2Compiler::compile_method (this=<optimized out>, env=0x7fffb5f5d5f0, target=0x7fff841f89b0, entry_bci=-1, install_code=true, directive=0x7ffff0217880)
    at /local/home/joshcao/jdk/jdk/src/hotspot/share/opto/c2compiler.cpp:119
#10 0x00007ffff58bb25c in CompileBroker::invoke_compiler_on_method (task=task@entry=0x7ffff03c5920) at /local/home/joshcao/jdk/jdk/src/hotspot/share/compiler/compileBroker.cpp:2281
#11 0x00007ffff58bc2a7 in CompileBroker::compiler_thread_loop () at /local/home/joshcao/jdk/jdk/src/hotspot/share/compiler/compileBroker.cpp:1942
#12 0x00007ffff5cfe551 in JavaThread::thread_main_inner (this=this@entry=0x7ffff03414c0) at /local/home/joshcao/jdk/jdk/src/hotspot/share/runtime/javaThread.cpp:720
#13 0x00007ffff5d04e98 in JavaThread::run (this=0x7ffff03414c0) at /local/home/joshcao/jdk/jdk/src/hotspot/share/runtime/javaThread.cpp:705
#14 0x00007ffff664e80f in Thread::call_run (this=this@entry=0x7ffff03414c0) at /local/home/joshcao/jdk/jdk/src/hotspot/share/runtime/thread.cpp:220
#15 0x00007ffff627fe8f in thread_native_entry (thread=0x7ffff03414c0) at /local/home/joshcao/jdk/jdk/src/hotspot/os/linux/os_linux.cpp:786
#16 0x00007ffff779e44b in start_thread () from /lib64/libpthread.so.0
#17 0x00007ffff72d552f in clone () from /lib64/libc.so.6
```