@caojoshua 2023-10-03
I reverted the patch to kill dead locals. The [GH action](https://github.com/navyxliu/jdk/actions/runs/6385809023/job/17333243056) ran previously with the original patch with two failures in GC tests.

* TestNoAllocationsInHRegions
* TestAllocHumongousFragment

They pass locally. I think they might be flaky. I want to see the test results