@navyxliu 2023-10-03
I am okay that we sideline kill_dead_locals() for time being. 
I slightly adjust your microbenchmark. I change its time unit form nano to micro and reduced the sized. 

For interval=2, I still feel that we expect to save norm allocation drop 50%. My reading is measly 26%. 
```
baseline
HashMapBench.replace:·gc.alloc.rate.norm                            2    1024  avgt   60  30720.228 ±     0.096    B/op
with PEA
HashMapBench.replace:·gc.alloc.rate.norm                            2    1024  avgt   60  22528.167 ±     0.076    B/op
```

```
diff --git a/test/micro/org/openjdk/bench/vm/compiler/pea/HashMapBench.java b/test/micro/org/openjdk/bench/vm/compiler/pea/HashMapBench.java
index bc149de08f3..b0903faa344 100644
--- a/test/micro/org/openjdk/bench/vm/compiler/pea/HashMapBench.java
+++ b/test/micro/org/openjdk/bench/vm/compiler/pea/HashMapBench.java
@@ -39,19 +39,19 @@
 import java.util.HashMap;

 @BenchmarkMode(Mode.AverageTime)
-@OutputTimeUnit(TimeUnit.NANOSECONDS)
+@OutputTimeUnit(TimeUnit.MICROSECONDS)
 @Warmup(iterations = 5, time = 1, timeUnit = TimeUnit.SECONDS)
-@Measurement(iterations = 10, time = 1, timeUnit = TimeUnit.SECONDS)
+@Measurement(iterations = 20, time = 1, timeUnit = TimeUnit.SECONDS)
 @State(Scope.Benchmark)
 @Fork(value = 3)
 public class HashMapBench {
-    @Param("100000")
+    @Param("1024")
     private int size;

-    @Param("2")
+    @Param("1", "2", "4")
     private int fillInterval;

-    HashMap<Integer, Object> map;
+    private HashMap<Integer, Object> map;

     @Setup
     public void setUp() {
```